move ExitConfig to earendil_topology crate
